Dont's use long here.
git-svn-id: http://svn.codehaus.org/jruby/trunk/jruby@8115 961051c9-f516-0410-bf72-c9f7e237a7b7

+++ src/org/jruby/RubyRange.java
@@ -519,7 +519,7 @@ public class RubyRange extends RubyObject {
         final RubyArray result = runtime.newArray(num);
         try {
             RubyEnumerable.callEach(runtime, context, this, new BlockCallback() {
-                long n = num;
+                int n = num;
                 public IRubyObject call(ThreadContext ctx, IRubyObject[] largs, Block blk) {
                     if (n-- <= 0) throw JumpException.SPECIAL_JUMP;
                     result.append(largs[0]);
